+++ src/main/resources/procedure.sql
@@ -1,5 +1,5 @@
 DELIMITER && 
-CREATE PROCEDURE proc_aumentaSalario(IN in_id INT)
+CREATE PROCEDURE proc_aumentaSalario(IN in_id FLOAT)
 READS SQL DATA
 BEGIN 
    UPDATE funcionario f SET f.salario = salario + (salario * (in_id/100));
